requirements.txt
fastapi==0.110.0
uvicorn==0.29.0

# Database
SQLAlchemy==2.0.30
psycopg2-binary==2.9.9

# Auth & Security
passlib[bcrypt]==1.7.4
python-jose==3.3.0

# Environment Variables
python-dotenv==1.0.1

# HTTP Requests
requests==2.31.0

# Fuzzy Matching
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0

# Caching
redis==5.0.4

# async requests
httpx==0.27.0


pydantic_settings==2.10.1
email-validator==2.2.0 
dnspython==2.7.0
cachetools==5.3.1  
